tests: Migrate HelpView tests to test_popups.
With the migration, this makes a minor rename of the test class to match
the tested class (ie. [Test]HelpView) from the previous TestHelpMenu.
