+++ tests/ui_tools/test_popups.py
@@ -4,7 +4,7 @@ import pytest
 from urwid import Columns, Text
 
 from zulipterminal.config.keys import keys_for_command
-from zulipterminal.ui_tools.views import AboutView, EditHistoryView
+from zulipterminal.ui_tools.views import AboutView, EditHistoryView, HelpView
 from zulipterminal.version import MINIMUM_SUPPORTED_SERVER_VERSION, ZT_VERSION
 
 
@@ -244,3 +244,34 @@ class TestEditHistoryView:
         return_value = EditHistoryView._get_author_prefix(snapshot, tag)
 
         assert return_value == expected_author_prefix
+
+
+class TestHelpView:
+    @pytest.fixture(autouse=True)
+    def mock_external_classes(self, mocker, monkeypatch):
+        self.controller = mocker.Mock()
+        mocker.patch.object(self.controller, 'maximum_popup_dimensions',
+                            return_value=(64, 64))
+        mocker.patch(VIEWS + ".urwid.SimpleFocusListWalker", return_value=[])
+        self.help_view = HelpView(self.controller, 'Help Menu')
+
+    def test_keypress_any_key(self, widget_size):
+        key = "a"
+        size = widget_size(self.help_view)
+        self.help_view.keypress(size, key)
+        assert not self.controller.exit_popup.called
+
+    @pytest.mark.parametrize('key', {*keys_for_command('GO_BACK'),
+                                     *keys_for_command('HELP')})
+    def test_keypress_exit_popup(self, key, widget_size):
+        size = widget_size(self.help_view)
+        self.help_view.keypress(size, key)
+        assert self.controller.exit_popup.called
+
+    def test_keypress_navigation(self, mocker, widget_size,
+                                 navigation_key_expected_key_pair):
+        key, expected_key = navigation_key_expected_key_pair
+        size = widget_size(self.help_view)
+        super_keypress = mocker.patch(VIEWS + '.urwid.ListBox.keypress')
+        self.help_view.keypress(size, key)
+        super_keypress.assert_called_once_with(size, expected_key)
